+++ Corruptor_Settings.h
@@ -14,7 +14,7 @@ class Corruptor_Settings : public QDialog
     Q_OBJECT
 
 public:
-    Corruptor_Settings(QWidget *parent, Settings *settings, int fileSize);
+    Corruptor_Settings(QWidget *parent, Settings *settings, qint64 fileSize);
     ~Corruptor_Settings();
 
 private slots:
@@ -37,10 +37,10 @@ private:
     void Set_Add_Mode(bool add);
     void Set_Shift_Left_Mode(bool shift);
     void Set_Replace_Mode(bool replace);
-    void Load_Settings(int fileSize);
+    void Load_Settings(qint64 fileSize);
     Ui::Corruptor_Settings *ui;
     Settings *settings;
-    int fileSize;
+    qint64 fileSize;
 };
 
 #endif // CORRUPTOR_SETTINGS_H
